# docs/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Eazy Recycling Backend Service
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /waste-streams/{wasteStreamNumber}/concept:
    put:
      tags:
      - waste-stream-controller
      operationId: update_1
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteStreamRequest"
        required: true
      responses:
        "204":
          description: No Content
  /waste-streams/{wasteStreamNumber}/active:
    put:
      tags:
      - waste-stream-controller
      operationId: updateAndValidate
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteStreamRequest"
        required: true
      responses:
        "204":
          description: No Content
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WasteStreamValidationResponse"
...
  /transport/waste/{id}:
    put:
      tags:
      - waste-transport-controller
      operationId: updateWasteTransport
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteTransportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateWasteTransportResponse"
...
  /waste-streams/concept:
    post:
      tags:
      - waste-stream-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteStreamRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateWasteStreamResponse"
  /waste-streams/active:
    post:
      tags:
      - waste-stream-controller
      operationId: createAndValidate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteStreamRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WasteStreamValidationResponse"
...
  /waste-streams:
    get:
      tags:
      - waste-stream-controller
      operationId: getWasteStreams
      parameters:
      - name: consignor
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteStreamListView"
  /waste-streams/{wasteStreamNumber}:
    get:
      tags:
      - waste-stream-controller
      operationId: getWasteStreamByNumber
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WasteStreamDetailView"
    delete:
      tags:
      - waste-stream-controller
      operationId: delete_1
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      responses:
        "204":
          description: No Content
  /waste-streams/{wasteStreamNumber}/weight-tickets:
    get:
      tags:
      - waste-stream-controller
      operationId: getWeightTicketsByWasteStream
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightTicketsByWasteStreamView"
  /waste-streams/{wasteStreamNumber}/compatible:
    get:
      tags:
      - waste-stream-controller
      operationId: findCompatibleWasteStreams
      parameters:
      - name: wasteStreamNumber
        in: path
        required: true
        schema:
          type: string
          pattern: "^[0-9]{12}$"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteStreamListView"
...
components:
  schemas:
    CompanyConsignor:
      allOf:
      - $ref: "#/components/schemas/ConsignorRequest"
      - type: object
        properties:
          companyId:
            type: string
            format: uuid
      required:
      - companyId
...
    DutchAddressRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
      - type: object
        properties:
          streetName:
            type: string
            minLength: 1
          buildingNumber:
            type: string
            minLength: 1
          buildingNumberAddition:
            type: string
          postalCode:
            type: string
            minLength: 1
            pattern: "^[1-9][0-9]{3}\\s?[A-Za-z]{2}$"
          city:
            type: string
            minLength: 1
          country:
            type: string
            minLength: 1
      required:
      - buildingNumber
      - city
      - country
      - postalCode
      - streetName
    NoPickupLocationRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
    PersonConsignor:
      allOf:
      - $ref: "#/components/schemas/ConsignorRequest"
    PickupCompanyRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
      - type: object
        properties:
          companyId:
            type: string
            format: uuid
      required:
      - companyId
...
    ProjectLocationRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
      - type: object
        properties:
          id:
            type: string
            format: uuid
          companyId:
            type: string
            format: uuid
          streetName:
            type: string
            minLength: 1
          buildingNumber:
            type: string
            minLength: 1
          buildingNumberAddition:
            type: string
          postalCode:
            type: string
            minLength: 1
            pattern: "^[1-9][0-9]{3}\\s?[A-Za-z]{2}$"
          city:
            type: string
            minLength: 1
          country:
            type: string
            minLength: 1
      required:
      - buildingNumber
      - city
      - companyId
      - country
      - id
      - postalCode
      - streetName
    ProximityDescriptionRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
      - type: object
        properties:
          description:
            type: string
            maxLength: 200
            minLength: 0
          postalCodeDigits:
            type: string
            minLength: 1
            pattern: "^[0-9]{4}$"
          city:
            type: string
            maxLength: 24
            minLength: 0
          country:
            type: string
            maxLength: 43
            minLength: 0
      required:
      - city
      - country
      - description
      - postalCodeDigits
...
    WasteStreamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        euralCode:
          type: string
          minLength: 1
        processingMethodCode:
          type: string
          minLength: 1
        collectionType:
          type: string
        pickupLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressRequest"
          - $ref: "#/components/schemas/NoPickupLocationRequest"
          - $ref: "#/components/schemas/PickupCompanyRequest"
          - $ref: "#/components/schemas/ProjectLocationRequest"
          - $ref: "#/components/schemas/ProximityDescriptionRequest"
        processorPartyId:
          type: string
          pattern: "^[0-9]{5}$"
        consignorParty:
          oneOf:
          - $ref: "#/components/schemas/CompanyConsignor"
          - $ref: "#/components/schemas/PersonConsignor"
        consignorClassification:
          type: integer
          format: int32
          maximum: 4
          minimum: 1
        pickupParty:
          type: string
          format: uuid
        dealerParty:
          type: string
          format: uuid
        collectorParty:
          type: string
          format: uuid
        brokerParty:
          type: string
          format: uuid
        catalogItemId:
          type: string
          format: uuid
      required:
      - collectionType
      - consignorClassification
      - consignorParty
      - euralCode
      - name
      - pickupLocation
      - pickupParty
      - processingMethodCode
      - processorPartyId
...
    ValidationErrorResponse:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
      required:
      - code
      - description
    ValidationRequestDataResponse:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        routeCollection:
          type: string
        collectorsScheme:
          type: string
        consignor:
          $ref: "#/components/schemas/ConsignorDataResponse"
        pickupLocation:
          $ref: "#/components/schemas/PickupLocationDataResponse"
        deliveryLocation:
          type: string
        consignorParty:
          $ref: "#/components/schemas/CompanyDataResponse"
        collectorParty:
          $ref: "#/components/schemas/CompanyDataResponse"
        dealerParty:
          $ref: "#/components/schemas/CompanyDataResponse"
        brokerParty:
          $ref: "#/components/schemas/CompanyDataResponse"
        wasteCode:
          type: string
        wasteName:
          type: string
        processingMethod:
          type: string
      required:
      - wasteStreamNumber
    WasteStreamValidationResponse:
      type: object
      properties:
        isValid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorResponse"
        requestData:
          $ref: "#/components/schemas/ValidationRequestDataResponse"
      required:
      - errors
      - isValid
...
    GoodsRequest:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        weight:
          type: number
          format: double
          minimum: 0
        unit:
          type: string
          minLength: 1
        quantity:
          type: integer
          format: int32
          minimum: 0
      required:
      - quantity
      - unit
      - wasteStreamNumber
      - weight
    WasteTransportRequest:
      type: object
      properties:
        pickupDateTime:
          type: string
          format: date-time
        deliveryDateTime:
          type: string
          format: date-time
        containerOperation:
          type: string
          enum:
          - EXCHANGE
          - PICKUP
          - EMPTY
          - DELIVERY
        transportType:
          type: string
          enum:
          - CONTAINER
          - WASTE
        driverId:
          type: string
          format: uuid
        carrierPartyId:
          type: string
          format: uuid
        truckId:
          type: string
        containerId:
          type: string
        note:
          type: string
        goods:
          type: array
          items:
            $ref: "#/components/schemas/GoodsRequest"
      required:
      - carrierPartyId
      - containerOperation
      - goods
      - note
      - pickupDateTime
      - transportType
    UpdateWasteTransportResponse:
      type: object
      properties:
        transportId:
          type: string
          format: uuid
        status:
          type: string
      required:
      - status
      - transportId
...
    DutchAddressView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
      - type: object
        properties:
          streetName:
            type: string
          postalCode:
            type: string
          buildingNumber:
            type: string
          buildingNumberAddition:
            type: string
          city:
            type: string
          country:
            type: string
      required:
      - buildingNumber
      - city
      - country
      - postalCode
      - streetName
...
    CreateWasteStreamResponse:
      type: object
      properties:
        wasteStreamNumber:
          type: string
      required:
      - wasteStreamNumber
...
    LocalDateTime:
      type: object
      properties:
        value:
          type: string
          format: date-time
          writeOnly: true
        time:
          $ref: "#/components/schemas/LocalTime"
        year:
          type: integer
          format: int32
        dayOfMonth:
          type: integer
          format: int32
          deprecated: true
          description: Use the 'day' property instead
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        month:
          type: string
          enum:
          - JANUARY
          - FEBRUARY
          - MARCH
          - APRIL
          - MAY
          - JUNE
          - JULY
          - AUGUST
          - SEPTEMBER
          - OCTOBER
          - NOVEMBER
          - DECEMBER
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        dayOfYear:
          type: integer
          format: int32
        date:
          $ref: "#/components/schemas/LocalDate"
        day:
          type: integer
          format: int32
        value$kotlinx_datetime:
          type: string
          format: date-time
        monthNumber:
          type: integer
          format: int32
          deprecated: true
          description: Use the 'month' property instead
        nanosecond:
          type: integer
          format: int32
      required:
      - date
      - day
      - dayOfMonth
      - dayOfWeek
      - dayOfYear
      - hour
      - minute
      - month
      - monthNumber
      - nanosecond
      - second
      - time
      - value$kotlinx_datetime
      - year
...
    WasteStreamListView:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        wasteName:
          type: string
        euralCode:
          type: string
        processingMethodCode:
          type: string
        consignorPartyName:
          type: string
        consignorPartyId:
          type: string
          format: uuid
        pickupLocation:
          type: string
        deliveryLocation:
          type: string
        status:
          type: string
        lastActivityAt:
          $ref: "#/components/schemas/LocalDateTime"
        isEditable:
          type: boolean
      required:
      - consignorPartyId
      - consignorPartyName
      - deliveryLocation
      - euralCode
      - isEditable
      - lastActivityAt
      - pickupLocation
      - processingMethodCode
      - status
      - wasteName
      - wasteStreamNumber
...
    WasteStreamDetailView:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        wasteType:
          $ref: "#/components/schemas/WasteTypeView"
        collectionType:
          type: string
        pickupLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressView"
          - $ref: "#/components/schemas/NoPickupView"
          - $ref: "#/components/schemas/PickupCompanyView"
          - $ref: "#/components/schemas/ProjectLocationView"
          - $ref: "#/components/schemas/ProximityDescriptionView"
        deliveryLocation:
          $ref: "#/components/schemas/DeliveryLocationView"
        consignorParty:
          oneOf:
          - $ref: "#/components/schemas/CompanyConsignorView"
          - $ref: "#/components/schemas/PersonConsignorView"
        consignorClassification:
          type: integer
          format: int32
        pickupParty:
          $ref: "#/components/schemas/CompanyView"
        dealerParty:
          $ref: "#/components/schemas/CompanyView"
        collectorParty:
          $ref: "#/components/schemas/CompanyView"
        brokerParty:
          $ref: "#/components/schemas/CompanyView"
        catalogItemId:
          type: string
          format: uuid
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
      required:
      - collectionType
      - consignorClassification
      - consignorParty
      - deliveryLocation
      - pickupParty
      - status
      - wasteStreamNumber
      - wasteType
    WasteTypeView:
      type: object
      properties:
        name:
          type: string
        euralCode:
          $ref: "#/components/schemas/EuralCodeView"
        processingMethod:
          $ref: "#/components/schemas/ProcessingMethodView"
      required:
      - euralCode
      - name
      - processingMethod
    WeightTicketsByWasteStreamView:
      type: object
      properties:
        weightTicketNumber:
          type: integer
          format: int64
        weightedAt:
          $ref: "#/components/schemas/LocalDateTime"
        amount:
          type: number
        createdBy:
          type: string
      required:
      - amount
      - weightTicketNumber
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT